Harmonizing type definitions and names

+++ src/clients/cs_addFileInfo.cpp
@@ -75,9 +75,9 @@ int main(int argc, char *argv[])
       return -3;
     }
 
-    unsigned long long startTime = getTime();
+    UInt64 startTime = getTime();
     int result = service->addFileInfo(sessionId,info);
-    unsigned long long endTime = getTime();
+    UInt64 endTime = getTime();
 
     if (result != 0)
     {
